Add query rendering test for filters on nested derived metrics
We have received a report of filters failing to render when they
are specified in the metric definition on a derived metric with
filters specified that is sourced from input derived metrics with
filters. Only the input filters are rendered.

This test illustrates the problem - we are, indeed, failing to
render the filter specified on the outer-most derived metric.